DataViews Extensibility: Allow unregistering the duplicate template part action
